Build the att-doc-latences figure
Have the top-level makefile build the included figure from
the R plotting code and data. This should allow a fresh
checkout of the paper to tex with a single `make` command.

Currently fails because data.csv isn't included in the repository,
or computed from another step. Usually it's good to have non-
deterministic data in the repo itself for consistency.